Extract rich mode account management
- Move rich account and token balance editing into its own component
- Normalize imported rich account balances and expand coverage
