+++ test/tests/richTokens.test.ts
@@ -2,7 +2,7 @@ import assert from 'node:assert'
 import { describe, test } from 'node:test'
 import type { IEthereumClientService } from '../../app/ts/simulation/services/EthereumClientService.js'
 import { MAKE_YOU_RICH_TRANSACTION } from '../../app/ts/utils/constants.js'
-import { addRichAccountBalanceOverrides, discoverErc1155BalanceStorage, discoverErc20BalanceStorageSlot, getDefaultRichTokenAmount, getErc1155BalanceStorageKey, getErc20BalanceStorageKey, getMatchingRichTokenOptions, getRichTokenOptions, MAX_RICH_TOKEN_AMOUNT, parseRichTokenAmountInput, verifyErc1155BalanceStorageSlot, verifyErc20BalanceStorageSlot } from '../../app/ts/utils/richTokens.js'
+import { addRichAccountBalanceOverrides, discoverErc1155BalanceStorage, discoverErc20BalanceStorageSlot, getDefaultRichTokenAmount, getErc1155BalanceStorageKey, getErc20BalanceStorageKey, getMatchingRichTokenOptions, getRichTokenOptions, MAX_RICH_TOKEN_AMOUNT, normalizeRichAccountBalances, parseRichTokenAmountInput, verifyErc1155BalanceStorageSlot, verifyErc20BalanceStorageSlot } from '../../app/ts/utils/richTokens.js'
 import { addressString, bigintToUint8Array, bytes32String } from '../../app/ts/utils/bigint.js'
 
 const owner = 0x1111111111111111111111111111111111111111n
@@ -56,6 +56,58 @@ describe('rich token support', () => {
 		const firstOption = options[0]
 		if (firstOption === undefined) throw new Error('Expected a rich token option')
 		assert.deepEqual(getMatchingRichTokenOptions([{ ...firstOption, tokenType: 'ERC1155', tokenId: 42n, symbol: 'ITEM' }], 'ITEM #42').map((option) => option.tokenId), [42n])
+	})
+
+	test('indexes large address books while preserving their original token order', () => {
+		const addressBookEntries = Array.from({ length: 5_000 }, (_, index) => ({
+			type: 'ERC20' as const,
+			name: `Indexed Token ${ index.toString() }`,
+			address: 0x1000n + BigInt(index),
+			symbol: `IDX${ index.toString() }`,
+			decimals: 18n,
+			entrySource: 'User' as const,
+			chainId: 1n,
+		}))
+
+		const options = getRichTokenOptions(1n, [], addressBookEntries)
+
+		assert.equal(options.length, addressBookEntries.length)
+		assert.equal(options[0]?.symbol, 'IDX0')
+		assert.equal(options.at(-1)?.symbol, 'IDX4999')
+	})
+
+	test('normalizes duplicate account profiles and token balances with latest values', () => {
+		const secondTokenAddress = tokenAddress + 1n
+		const normalized = normalizeRichAccountBalances([
+			{
+				chainId: 1n,
+				address: owner,
+				nativeAmount: 10n,
+				tokenBalances: [
+					{ tokenAddress, tokenId: undefined, amount: 100n },
+					{ tokenAddress: secondTokenAddress, tokenId: undefined, amount: 200n },
+				],
+			},
+			{
+				chainId: 1n,
+				address: owner,
+				nativeAmount: 20n,
+				tokenBalances: [
+					{ tokenAddress, tokenId: undefined, amount: 300n },
+					{ tokenAddress, tokenId: undefined, amount: 400n },
+				],
+			},
+		])
+
+		assert.deepEqual(normalized, [{
+			chainId: 1n,
+			address: owner,
+			nativeAmount: 20n,
+			tokenBalances: [
+				{ tokenAddress, tokenId: undefined, amount: 400n },
+				{ tokenAddress: secondTokenAddress, tokenId: undefined, amount: 200n },
+			],
+		}])
 	})
 
 	test('offers ERC-20 address book entries without requiring a preset', () => {
